change structure and add readme
